# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Гармония (Harmony) API
  version: 1.0.0
  summary: REST and realtime API for the Harmony social network.
  description: |
    # Гармония API

    A social network API. All responses share one envelope so a client has a
    single code path for success and failure:

    ```json
    { "ok": true,  "data": { }, "meta": { "count": 15, "has_more": true, "next_cursor": "eyJ..." } }
    { "ok": false, "error": { "code": "not_found", "message": "Пост не найден.", "request_id": "..." } }
    ```

    ## Authentication

    After `POST /auth/login` the API issues two cookies:

    | Cookie | HttpOnly | Purpose |
    |---|---|---|
    | `harmony_access` | yes | Access token (JWT) |
    | `harmony_refresh` | yes | Refresh token, scoped to `/api/v1/auth` |
    | `harmony_csrf` | **no** | CSRF token for the double-submit pattern |

    The access cookie is `HttpOnly` so that an XSS payload cannot read it.
    The CSRF cookie is deliberately readable, because the client must echo it.

    **Every unsafe method** (`POST`, `PATCH`, `PUT`, `DELETE`) must send the CSRF
    token in the `X-CSRF-Token` header, matching the cookie. A correct `Origin`
    or `Referer` is also accepted, which keeps non-browser clients working.

    Non-browser clients may instead send `Authorization: Bearer <access_token>`.

    ## Resources

    Resources are addressed by an opaque 32-character `public_id`, never by the
...
    bugs easier to exploit.

    ## Pagination

...

...

...

...

...

...

...

  contact:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /feed:
    get:
      tags: [Posts]
      summary: The feed
      security: []
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [latest, following]
            default: latest
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Posts, newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedPosts" }
  /posts/search:
    get:
      tags: [Posts]
      summary: Search posts
      description: Full-text on PostgreSQL, with a `LIKE` fallback for other dialects.
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2 }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Matching posts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedPosts" }
  /posts:
    post:
      tags: [Posts]
      summary: Publish a post
      description: |
        Text is sanitised and screened before storage. Images are uploaded
        separately to `/uploads/images` and referenced by `storage_key`; the
        server looks up the real metadata, so a client cannot attach an
        arbitrary URL or mislabel a MIME type.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body: { type: string, maxLength: 5000 }
                visibility:
                  type: string
                  enum: [public, followers, private]
                  default: public
                media:
                  type: array
                  maxItems: 4
                  items: { $ref: "#/components/schemas/MediaReference" }
      responses:
        "201":
          description: Published
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      post: { $ref: "#/components/schemas/Post" }
                    additionalProperties: true
        "202":
          description: Held for moderator review
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      status: { type: string, const: under_review }
                      message: { type: string }
        "422": { description: Empty post or validation failure }
        "429": { description: Post or spam limit reached }
        "451": { description: Rejected by content policy }
  /posts/{post_id}:
    get:
      tags: [Posts]
      summary: Read a post
      security: []
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      post: { $ref: "#/components/schemas/Post" }
                    additionalProperties: true
        "404": { description: "Not found, or not visible to the caller" }
    patch:
      tags: [Posts]
      summary: Edit a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body: { type: string, maxLength: 5000 }
                visibility: { type: string, enum: [public, followers, private] }
      responses:
        "200": { description: Updated }
        "403": { description: Not the author }
        "404": { description: Not found }
    delete:
      tags: [Posts]
      summary: Delete a post
      description: "Soft delete: the row is retained as a tombstone so comments and moderation history stay coherent."
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204": { description: Deleted }
        "403": { description: Not the author }
  /posts/{post_id}/restore:
    post:
      tags: [Posts]
      summary: Restore a deleted post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200": { description: Restored }
        "409": { description: Post was not deleted }
  /posts/{post_id}/reactions:
    post:
      tags: [Posts]
      summary: Like or unlike
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [like, unlike], default: like }
      responses:
        "200":
          description: New state
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      liked: { type: boolean }
                      likes_count: { type: integer }
    get:
      tags: [Posts]
      summary: Who liked a post
      security: []
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Reactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      reactions:
                        type: array
                        items:
                          type: object
                          properties:
                            type: { type: string }
                            user: { $ref: "#/components/schemas/User" }
                            created_at: { type: string, format: date-time }
  /posts/{post_id}/media:
    post:
      tags: [Posts]
      summary: Attach previously uploaded images
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                media:
                  type: array
                  items: { $ref: "#/components/schemas/MediaReference" }
      responses:
        "201": { description: Attached }
        "422": { description: Unknown or already-claimed files }

  # ------------------------------------------------------------ Comments ---
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: harmony_access
      description: |
        Access token set by `/auth/login`. Send `X-CSRF-Token` on unsafe methods.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: For non-browser clients. Exempt from CSRF.
  parameters:
...
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9a-f]{32}$' }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from `meta.next_cursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 15, minimum: 1, maximum: 100 }
  schemas:
...
    PageMeta:
      type: object
      properties:
        count: { type: integer }
        has_more: { type: boolean }
        next_cursor: { type: [string, "null"] }
        total: { type: integer }
...
    User:
      type: object
      description: |
        The public projection. Never includes the email unless the target has
        opted in, nor any moderation state, counter or token.
      properties:
        public_id: { type: string }
        username: { type: string }
        display_name: { type: [string, "null"] }
        initials: { type: string }
        bio: { type: [string, "null"] }
        avatar_url: { type: [string, "null"] }
        avatar_color: { type: string }
        location: { type: [string, "null"] }
        website: { type: [string, "null"] }
        pronouns: { type: [string, "null"] }
        role: { type: string, enum: [member, moderator, admin] }
        is_verified: { type: boolean, description: Active account with a confirmed email. }
        posts_count: { type: integer }
        comments_count: { type: integer }
        followers_count: { type: integer }
        following_count: { type: integer }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        email: { type: string, description: Present only if the target enables it. }
        profile_visibility: { type: string }
        is_self: { type: boolean }
...
    Post:
      type: object
      properties:
        id: { type: string }
        author: { $ref: "#/components/schemas/User" }
        body: { type: string, description: Plain text, always. The API never returns HTML. }
        spans:
          type: array
          description: |
            Typed text segments (link / mention / hashtag / text). Rendering
            these as DOM nodes keeps user content out of the HTML parser.
          items:
            type: object
            properties:
              type: { type: string, enum: [text, link, mention, hashtag] }
              text: { type: string }
              url: { type: string }
              username: { type: string }
              tag: { type: string }
        media:
          type: array
          items: { $ref: "#/components/schemas/Media" }
        visibility: { type: string, enum: [public, followers, private] }
        status: { type: string, enum: [published, hidden, under_review, removed, archived] }
        likes_count: { type: integer }
        comments_count: { type: integer }
        viewer_has_liked: { type: boolean }
        is_owner: { type: boolean }
        can_edit: { type: boolean }
        can_delete: { type: boolean }
        can_report: { type: boolean }
        created_at: { type: string, format: date-time }
        edited_at: { type: [string, "null"], format: date-time }
        status_reason: { type: string }
    Media:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
        thumbnail_url: { type: string }
        width: { type: integer }
        height: { type: integer }
        byte_size: { type: integer }
        mime_type: { type: string }
        alt_text: { type: [string, "null"] }
        storage_key: { type: string }
    MediaReference:
      type: object
      required: [storage_key]
      properties:
        storage_key:
          type: string
          description: Returned by `/uploads/images`. The server resolves the real metadata itself.
        alt_text: { type: string, maxLength: 240 }
...
    PaginatedPosts:
      type: object
      properties:
        ok: { type: boolean }
        data:
          type: array
          items: { $ref: "#/components/schemas/Post" }
        meta: { $ref: "#/components/schemas/PageMeta" }
...